+++ src/property/error.rs
@@ -1,3 +1,5 @@
+use std::{error::Error, fmt::Display};
+
 use super::{PropertyValueKind, UiProperty};
 
 /// A typed property/value mismatch.
@@ -10,3 +12,15 @@ pub struct PropertyValueError {
     /// The actual value kind.
     pub actual: PropertyValueKind,
 }
+
+impl Display for PropertyValueError {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(
+            f,
+            "Property value error: expected {}, got {}",
+            self.expected, self.actual
+        )
+    }
+}
+
+impl Error for PropertyValueError {}